+++ pkgs/inheritCargoArtifacts.nix
@@ -1,12 +1,8 @@
 { makeSetupHook
-, pkgsBuildBuild
 }:
 
 makeSetupHook
 {
   name = "inheritCargoArtifactsHook";
-  substitutions = {
-    zstd = "${pkgsBuildBuild.zstd}/bin/zstd";
-  };
 } ./inheritCargoArtifactsHook.sh