+++ pkgs/installCargoArtifacts.nix
@@ -1,11 +1,7 @@
 { makeSetupHook
-, pkgsBuildBuild
 }:
 
 makeSetupHook
 {
   name = "installCargoArtifactsHook";
-  substitutions = {
-    zstd = "${pkgsBuildBuild.zstd}/bin/zstd";
-  };
 } ./installCargoArtifactsHook.sh