+++ lib/cargoTest.nix
@@ -1,0 +1,22 @@
+{ mkCargoDerivation
+}:
+
+{ cargoArtifacts
+, cargoExtraArgs ? ""
+, cargoTestExtraArgs ? ""
+, ...
+}@origArgs:
+let
+  args = (builtins.removeAttrs origArgs [
+    "cargoExtraArgs"
+    "cargoTestExtraArgs"
+  ]);
+in
+mkCargoDerivation (args // {
+  inherit cargoArtifacts;
+  doCheck = args.doCheck or true;
+
+  pnameSuffix = "-test";
+  buildPhaseCargoCommand = "";
+  checkPhaseCargoCommand = "cargoWithProfile test ${cargoExtraArgs} ${cargoTestExtraArgs}";
+})