+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `crateNameFromCargoToml`) if they are not specified
 * `mkCargoDerivation` now defaults to an empty `checkPhaseCargoCommand` if not
   specified
+* `cargoAudit` now delegates to `mkCargoDerivation` instead of `cargoBuild`
 * `cargoDoc` now delegates to `mkCargoDerivation` instead of `cargoBuild`
 * `cargoFmt` now delegates to `mkCargoDerivation` instead of `cargoBuild`