+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * `mkCargoDerivation` now defaults to an empty `checkPhaseCargoCommand` if not
   specified
 * `cargoAudit` now delegates to `mkCargoDerivation` instead of `cargoBuild`
+* `cargoClippy` now delegates to `mkCargoDerivation` instead of `cargoBuild`
 * `cargoDoc` now delegates to `mkCargoDerivation` instead of `cargoBuild`
 * `cargoFmt` now delegates to `mkCargoDerivation` instead of `cargoBuild`
 * `cargoTarpaulin` now delegates to `mkCargoDerivation` instead of `cargoBuild`